add length( EXPR ) function
Tries to DWIM as much as possible, though I think numbers are going to
be a bit weird (will get the number of base-10 digits no matter what, I
suspect, but that's not a use-case I need right now).

I won't be doing a scalar() alias, but I might consider a count() alias.

Fixes #28