add sort(EXPR) function to sort by string
This does not support numeric sorting. Like group_by(EXPR), this buffers
the output. Like all functions that use scope, this probably will not
work in assignment.

I'm not sure how I want to do numeric sorting. Probably as a first
argument of some kind. Maybe if sort took two EXPRs and we added the <=>
and cmp binary operators, and the magic $a and $b variables...

Fixes #12