fix default format module in csv format
use_module() from Module::Runtime has a prototype that enforces scalar
context on an array being passed in, so we can't use an array for the
parameters.

Since the CSV module uses state to keep the heading line, we have to
build new closures for every instance. This is not a good way to do this
and we need to come up with a better one...