add error checking when preparing or executing SQL
Previously, a warning would be printed about a bad prepare, and then the
code would continue on as nothing happened, finally dying when trying to
run the execute() method on an undef value.

Now we give the error message, without the usage information that
usually accompanies our error messages (if they made it to a SQL error,
they know how to use the tool).

Fixes #92