add initial ygrok script
ygrok is a way to convert lines of plaintext into YAML documents that
can be futher used by Yertl (converted to CSV or JSON with yto, or put
into a database with ysql).

This is extremely preliminary, and there are lots of future development
to do, but this is the start.

Refs #105