fix "prototype mismatch" from Parse::RecDescent
Parse::RecDescent destroys the entire namespace it creates, but it does
so by first breaking potential cycles by saving a blank subref to the
stash. If that sub it's overwriting had a prototype, we get a warning.

We shouldn't do all this in the parsing anyway, we should build a tree
to comb through later, but for now, not importing the prototyped subs
fixes this.

Fixes #103