fix yq filter on empty document returning nonempty
'empty' is a special value that is propagated up and ignored when
printing out the results. But when we try to additionally filter
something that could return empty, the empty was changed to undef,
which gets printed. Meaning these two filters can differ in the number
of results returned:

    select( .foo == 1 )
    select( .foo == 1 ) | .bar

Now, any time a selection happens on an empty document, the result is
also empty, which will be ignored correctly.

Fixes #132