test that files do not need separators
A while ago, when we would try to parse multiple files that didn't start
with the standard `---` YAML document separator, we would get an error:

    YAML Error: Expected separator '---'

Since I can't reproduce it anymore, I'm going to assume it was fixed
when we switched over to using Format modules.

We are now at least testing this case to ensure it doesn't happen again.

Fixes #45