add better error when inserting deep reference
When we give a hashref or arrayref as a column value to SQL::Abstract,
it gives an ugly warning and the resulting query is likely to fail. Now
we pre-check that the data is sane before we insert, so that the error
doesn't happen and users are instead pointed towards a better solution.