allow function calls on either side of binary ops
Now we can assign the output of a function to a filter, or compare the
output of a function with a filter. This was a limitation of the
grammar, so we did a lot of changes to how we're using the regular
expressions (thanks Damien Conway for demonstrating this technique with
his PPR module).

Fixes #29