fix output from yto not having all documents
We need to flush the output before we can exit our program. Autoflush
isn't a perfect solution, but it works for now.

We'll have a better solution when we refactor looping over a bunch of
files asynchronously but serially. Then we can use on_close and
close_when_empty on the output handle to ensure all the data is written.