fix reading partial yaml documents
Before parsing a YAML document, we need to make sure we've got
a complete one. The previous regex for doing this wasn't correct: It
allowed the end of the string no matter what, which might be the end of
the buffer (and load an incomplete document).

Now we loop over the YAML line-wise looking for the document marker, and
only consider the end of the string if we've reached the end of the
file.