fix format being lost after reconfiguring stream
The IO::Async model allows for the `configure` method to be called many
times to reconfigure the object. When that happens, we need to only
change the configuration that the user wants us to change.

It is probably best that default configuration go into the `_init`
method, but I'm not doing that at the moment. Let's see how this goes
for now...