add proof-of-concept transform class
This is a very good start. There is a lot of boilerplate we can solve
with some helper functions later.

It doesn't solve the problem of needing raw input for something like
"ygrok" to work, but right now I like the idea of "ygrok" being an I/O
protocol (like FilterStream will be when I understand
IO::Async::Protocol better).