add extra args to streams during inputseries
For `yfrom` to work, we need to be able to specify the format of the
input. Now we allow passing in a "format" to InputSeries which will then
be passed to the streams.

This requires that we delay creating the initial stream. For the maximum
delay (and thus maximum flexibility), we now delay until the InputSeries
object is added to the loop.