fix args to Import::Base becoming empty strings
What we really need to do when looping over an array and removing items
from it is to use the C-style for loop: The C-style for loop continually
re-evaluates its end condition, so if `$#_` changes, we compare against
the changed value. Also, while we're looping, the `redo` might end up
past the array, so we have to make sure not to redo in that case.